#85 fix: 콘솔 제거 및 코드 가독성 향상
